Changing serial commands to make the various ifdef options more comparable.
A few of the serial commands (gps, mag) were not the same for the different IFDEF options available. This changeset attempts to make these different options more equal so that a programmatic interface to the configuration doesn't need to know as much about the internal configuration in order to read consisten output. Zeros are printed, or dummy values read, instead of the real values where necessary.