Adds scale factors into the compass code, enabled for the HMC58xx magnetometer, with scaling set to 1.0 on each axis.
This changeset adds a magScale factor into the code to allow a scaling of each compass axis independently. Scaling occurs before the bias is added so the bias value should include the scale factor as well (if bias is calculated before scaling). No configuration code has been added to set the magScale factors, but memory addresses have been allocated next to the magBias factors as is done in the accelerometer and gyro code. Default values are 1,1,1 so no scaling occurs.

Changes in DataStorage.h cover the allocation of EEPROM space (and in ITG3200TempCalibration/DataStorage.h). Changes in AeroQuad.h set the default values, read from EEPROM and write to EEPROM.

The Compass.h interface and the HMC58xx interface files have been modified to include the scale factor, the other compass class have not been modified.

This changeset does not allow for the calibration or even setting of the scale factors. This depends on whether the code is standard or optional. If it's standard the configuration option 'M' should be used to set the Scaling along with the biases. If it's not standard, another case option needs to be found.