Update to mag/Compass code.
Integrates the magScale parameter a bit more closely into the code, and defines a new variable and function to access the scales+bias corrected data.

A new variable called measuredMag (3-vector) is defined that is currently a duplicate of measuredMagX, Y, Z because they are used elsewhere in the code. A new function called getMagnetometerData(byte axis) is defined to access this 3 vector with the same functionality of getMagnetometerRawData(byte axis).

A new user defined variable "UserMeasuredMag" is used to determine whether the rawData or 'measured' data is returned in serial com and OSDMenu. *except* in the case 'j' where
raw magnetometer data is always returned for calibration purposes.