Adds code to run the serial communication over XBEE or USB.
This changeset initializes USB and XBEE serial ports on the AeroQuad at the same time, and allows communication over either port. The board will respond on the same port
that received the command. There is a new serial option 'Q' that can be used to configure the XBEE (following page 27 of the XBEE-pro documentation for the commands) and everything is wrapped in a new define "MultipleSerialTelemetry" that is independent of the default option or the WirelessTelemetry option.