Some work on the test sketches
Notes on what I've done:
Where possible, I've put in a detection check for the sensor - if it's
assigned a vehicleState, I check this after initialization and if the
sensor was not found, the sketch says so and grinds to a halt. If it is
found, it says " xxxxx found" and continues on.
- I restructured the timing - some of the sketches are reading things
at 100Hz but trying to read serial monitor that fast is painful, so I
slowed down the serial comm to 25Hz in some cases.

PROBLEMS SO FAR:
- None of the accelerometer test sketches "work." The accelerometer on
my 9DOF works fine, I fly with it frequently and the sensor data looks
good in Configurator, but I can't get the test sketch to read
meaningful data. Both the ADXL345 and BMA180 are detected by the test
sketch but the data coming out is 0s and 9.81...

- HMC5883L - pitch and roll work but yaw stays stuck in one direction,
I'm not sure if this is a hardware problem on my end or if the test
sketch is messing up.

- sBUS - works except pitch/roll/yaw. No idea why.