Fixed some things
Removed 100Hz cameraControl

Reinstated USE_USB_SERIAL - users still need to comment this out to use
wireless telemetry. This needs to be fixed. The user should not have to
go into SerialMapping.h to change things to be able to use telemetry
(defining WirelessTelemetry in UserConfiguration.h DOES NOT disable
SerialUSB and DOES NOT enable serial data on Serial1 like it should).
I've looked at this stuff but it's a mess.

Fixed Test_ADXL345.h and Test_BMA180.h