Some AQ32 batch files
bl64k.img - kha's bootloader that clears EEPROM, etc. Necessary for
switching from aq32plus back to AQ, I think this should be included
with the software as using the webcompiler to build a bootloader does
not seem to work when switching from aq32plus to AQ.

flash_bl64k.bat - batch file to flash the above bootloader

buildAndFlash.bat - batch file that asks user if they want to build
Libmaple and/or AQ32 and builds what the user asks for. It then runs
dfu-util.exe to list connected DFU devices and asks the user if they'd
like to upload code.

This makes it so you don't have to use Configurator to compile code OR
type anything in a command prompt (other than "y" or "n" when it
prompts you). I use it as a streamlined way to build and upload when
testing things.
/tools/dfu-util.exe - Of course, this all depends on dfu-util.exe being
in the tools directory (FlashDFUSeF4.bat and ListDFUSeDevices.bat are
written expecting it there). I'm not sure why this isn't included when
all of the necessary build and flash batch files are. If you don't want
to include dfu-util.exe in the AeroQuad directory (it's free software,
but I suppose it could still be seen as distributing protected
software?) then just go ahead and ignore all of the batch files. Please
still pull in kha's bootloader if possible.